requirements.txt
asgiref==3.2.7
Django==2.0
django-autoslug==1.9.7
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-js-asset==1.0.0
djangorestframework==3.11.0
numpy==1.18.2
psycopg2==2.8.5
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
